requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-JWT-Extended==4.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
passlib==1.7.4
protobuf==3.20.3
PyJWT==2.7.0
Werkzeug==2.3.6
